Don't track paps for pods
This change removes pap tracking for people in capsules.  The main reason here is if a fleet dies and the pap job runs (which it will), the pap log will only show people in capsules instead of what they were originally in, which is not ideal.

Also, not sure anyone ever cares to track paps for someone who is in a capsule anyway if they had never joined in a real ship to start with.